Fix imports and references
