+++ src/plugins/db/utils.ts
@@ -1,28 +1,8 @@
 import {
-  sql,
   InterceptorType,
   FieldType,
   QueryResultRowColumnType,
 } from 'slonik';
-
-export const camelToSnakeCase = (str: string) =>
-  str.replace(/[A-Z]/g, letter => `_${letter.toLowerCase()}`);
-
-export type DateTime = Date;
-
-export type SnakeToCamel<S extends string> = S extends `${infer T}_${infer U}`
-  ? `${T}${Capitalize<SnakeToCamel<U>>}`
-  : S;
-
-export type CamelToSnake<S extends string> = S extends `${infer T}${infer U}`
-  ? `${T extends Capitalize<T> ? '_' : ''}${Lowercase<T>}${CamelToSnake<U>}`
-  : S;
-
-export const select = (columns: string[]) =>
-  sql.join(
-    columns.map(column => sql.identifier([column])),
-    sql`, `,
-  );
 
 export const transformNameInterceptors: () => InterceptorType = () => {
   const underscoreFieldRegex = /^[a-z0-9_]+$/;
