+++ src/queries/action-type-queries.ts
@@ -1,6 +1,6 @@
 import { sql } from 'slonik';
-import { DateTime } from '../utils';
-import { SnakeToCamel } from '../utils';
+import { DateTime } from './utils';
+import { SnakeToCamel } from './utils';
 
 class ActionTypeRow {
   id: number;
